+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Reactist follows [semantic versioning](https://semver.org/) and doesn't introduce breaking changes (API-wise) in minor or patch releases. However, the appearance of a component might change in a minor or patch release so keep an eye on redesigns and make sure your app still looks and feels like you expect it.
 
-# Next
+# v15.2.0
 
 -   [Feat] Enhance the `label` property in `CheckboxField` to be of type `React.ReactNode`